% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_cleaning.R
\name{clean_timeseries_RSEI}
\alias{clean_timeseries_RSEI}
\title{Get RSEI data that is consistent for a time series}
\usage{
clean_timeseries_RSEI(
  rsei = NA,
  chemicals = NA,
  facility = NA,
  starting_year = 2011
)
}
\arguments{
\item{rsei}{The dataset as read in from the Duke Box folder \code{raw/rsei}. No changes made, or else output isn't guaranteed, MUST be passed as a data.frame.}

\item{chemicals}{The dataset as read in from the Duke Box folder \code{raw/rsei/chemical_data_rsei_v2310.csv}. MUST be passed as a data.frame.}

\item{starting_year}{The year you want to start having a consistent dataset from. The function will only return data for the year you've inputted, but will drop certain data based on the year you input here.}
}
\value{
a dataset of RSEI data, with certain data filtered out based on your \code{starting_year} of choice.
}
\description{
Get RSEI data that is consistent for a time series
}
