% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_cleaning.R
\name{detach_all_packages}
\alias{detach_all_packages}
\title{A function to detach all packages}
\usage{
detach_all_packages()
}
\description{
Primarily useful for checking replication, or when switching between things you're working on and you're not using R Projects.
Credit to: https://stackoverflow.com/questions/7505547/detach-all-packages-while-working-in-r
}
