% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapefiles.R
\name{simplify_and_combine_shapefiles}
\alias{simplify_and_combine_shapefiles}
\title{Combining multiple shapefiles}
\usage{
simplify_and_combine_shapefiles(shape_paths, id, tol, topologyPreserve = T)
}
\arguments{
\item{id}{What column in the data to use as a UNIQUE id for the shapefiles. Must be unique values across all shapefiles}

\item{tol}{What your tolerance is for simplification, smaller number means more similar output and less of a change in object size. Larger number means less similar output, larger change in object size. More info in gSimplify docs.}

\item{topologyPreserve}{see gSimplify}

\item{shape}{a vector of paths to where your shapefiles you want to simplify and combine are}
}
\description{
A frequent data munging problem is simplifying and combining a bunch of shapefiles. This function takes a list of SPDF's, and returns you one large one, dealing with ID issues, and making sure that the data frame stays attached.
}
